add tests for users

+++ auth/src/app/services/role_service.py
@@ -1,6 +1,8 @@
+from http import HTTPStatus
 from uuid import UUID
 
 from ..db.database import AbstractDatabase, User, Role
+from ..exceptions import HTTPError
 
 database: AbstractDatabase
 
@@ -23,31 +25,46 @@ def add_role(name: str) -> Role:
 
 
 def delete_role(role_id: UUID):
+    if database.role_by_id(role_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="Role not found")
     database.delete_role(role_id)
 
 
 def update_role(role_id: UUID, new_name: str):
+    if database.role_by_id(role_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="Role not found")
     database.update_role(role_id, new_name)
 
 
 def get_role(role_id: UUID):
-    role = database.role_by_id(role_id)
+    if (role := database.role_by_id(role_id)) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="Role not found")
     return role.dict()
 
 
 def get_user_roles(user_id: UUID):
+    if database.user_by_id(user_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
     roles_list = database.get_user_roles(user_id)
     result = [role.dict() for role in roles_list]
     return result
 
 
 def add_user_role(user_id: UUID, role_id: UUID):
+    if database.user_by_id(user_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
+    if database.role_by_id(role_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="Role not found")
     roles_list = database.add_user_role(user_id, role_id)
     result = [role.dict() for role in roles_list]
     return result
 
 
 def delete_user_role(user_id: UUID, role_id: UUID):
+    if database.user_by_id(user_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
+    if database.role_by_id(role_id) is None:
+        raise HTTPError(status_code=HTTPStatus.NOT_FOUND, detail="Role not found")
     roles_list = database.delete_user_role(user_id, role_id)
     result = [role.dict() for role in roles_list]
     return result
